pyproject.toml
[project]
name = "twitch-eventsub-service"
version = "0.1.0"
description = "Minimal Twitch EventSub bridge service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.16.0",
  "asyncpg>=0.30.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.10.0",
  "python-dotenv>=1.1.0",
  "sqlalchemy[asyncio]>=2.0.40",
  "uvicorn[standard]>=0.35.0",
  "websockets>=15.0.1",
  "prompt-toolkit>=3.0.51",
]

[project.optional-dependencies]
test = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
]

[project.scripts]
twitch-eventsub-api = "app.main:run"
twitch-eventsub-cli = "app.cli:main"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
  "unit: fast deterministic unit tests",
  "integration: integration tests that use DB/app wiring",
]
addopts = "-ra -p no:cacheprovider"
norecursedirs = [".git", ".venv", ".tmp", ".tmp-pip", "logs", "pytest-cache-files-*"]

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 35
